requirements.txt
certifi==2020.6.20
cycler==0.10.0
dataclasses==0.6
filterpy==1.4.5
future==0.18.2
joblib==1.0.1
kiwisolver==1.3.0
matplotlib==3.3.2
numpy==1.19.2
opencv-python==4.4.0.44
pandas==1.2.3
Pillow==8.0.1
prettytable==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
scikit-learn==0.24.1
scipy==1.5.3
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.7.0
torchvision==0.8.1
tqdm==4.51.0
typing-extensions==3.7.4.3
wcwidth==0.2.5
websockets==8.1
